Add files via upload
q1: 0 pt. To answer this question, let’s look at the confusion matrix. But first, we need to remember that the matrix uses 0s and 1s for class labels instead of terms like ‘originated’ or ‘denied’. From the assignment, we know that “after using LabelEncoder, the class labels are 0s (originated) and 1s (denied).” Looking at the confusion matrix, the top row shows all the loans that actually have a class label of 0 (originated). So, the total number of originated loans is 2038 + 224, which is 2262. Out of these, the model correctly predicted 2038 loans as originated. So, the answer to this question is 2038 out of 2262, or 90.10%. q2: 0 pt. Looking at the confusion matrix, the left column shows all the loans that the model predicted to have a class label of 0 (originated). So, the total number of loans that were predicted to be originated is 2038 + 513, which is 2551. Out of these, 2038 loans were actually originated. So, the answer to this question is 2038 out of 2551, or 79.89%. You may have flipped the logic needed for questions 1 and 2. q3: 0 pt. To approve more loans, the model should return more 0s. By raising the threshold, more samples will be labeled as 0 and fewer samples will be labeled as 1. So, the answer is: they need to use a higher threshold. q4: 1 pt. q5: 2 pt. q6: 1.5 pt.
Reza Mousavi at Mon Jan 22, 2024 10:52am1/22/2024
Between Q1 and Q2, I decided to deduct 1 pt instead of 2 pts. Also, for Q3, I decided to deduct 0.75 pt instead of 1 pt.